+++ Magic/CHANGELOG.md
@@ -31,6 +31,7 @@
  - See: https://reference.elmakers.com/#generators and https://reference.elmakers.com/#populators
  - Allow modifying game rules of a world via `game_rules` config
  - Add `time` parameter to set a world to a specific time of day
+ - Add `time_distance` world parameter, can be used with `time` where both are ranges, and time will vary by distance from spawn
  - Add `title_delay` parameter to show players a title when they enter a world
  - Add `respawn_world` parameter to force players to respawn in a different world
  - Add `world_modification.transfer` map, to transfer players from one world to another automatically